refactor: 폴더명 firebase -> play_list 변경
