+++ lib/data/source/play_list/play_list_source_impl.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:oz_player/data/dto/play_list_dto.dart';
-import 'package:oz_player/data/source/firebase/play_list_source.dart';
+import 'package:oz_player/data/source/play_list/play_list_source.dart';
 
 class PlayListSourceImpl implements PlayListSource {
   final FirebaseFirestore _firestore;
